unified syntax  - interactor refactor
