+++ framework/snippets/src/interactor_tx/interactor_exec_deploy.rs
@@ -1,62 +1,50 @@
-use std::path::PathBuf;
-
 use multiversx_sc_scenario::{
-    api::StaticApi,
     multiversx_sc::{
         tuple_util::NestedTupleFlatten,
         types::{
-            AnnotatedValue, Code, DeployCall, FunctionCall, ManagedAddress, ManagedBuffer,
-            RHListExec, Tx, TxBaseWithEnv, TxCodeSource, TxCodeSourceSpecified, TxCodeValue, TxEnv,
-            TxFromSpecified, TxGas, TxPayment, TxToSpecified,
+            Code, DeployCall, RHListExec, Tx, TxBaseWithEnv, TxCodeValue, TxFromSpecified, TxGas,
+            TxPayment,
         },
     },
-    scenario::tx_to_step::{StepWrapper, TxToStep},
-    scenario_model::{AddressValue, BytesValue, ScCallStep, ScDeployStep, TxResponse},
-    ScenarioEnvExec, ScenarioTxEnv, ScenarioTxEnvData, ScenarioTxRun, ScenarioWorld,
+    scenario::tx_to_step::TxToStep,
+    scenario_model::{ScDeployStep, TxResponse},
+    ScenarioTxEnvData,
 };
 
-use crate::{Interactor, InteractorPrepareAsync};
+use crate::Interactor;
 
-use super::InteractorEnvExec;
-
-pub struct InteractorDeployStep<'w, RH>
-where
-    RH: RHListExec<TxResponse, InteractorEnvExec<'w>>,
-    RH::ListReturns: NestedTupleFlatten,
-{
-    step_wrapper: StepWrapper<InteractorEnvExec<'w>, ScDeployStep, RH>,
-}
+use super::{InteractorExecEnv, InteractorExecStep, InteractorPrepareAsync};
 
 impl<'w, From, Payment, Gas, CodeValue, RH> InteractorPrepareAsync
     for Tx<
-        InteractorEnvExec<'w>,
+        InteractorExecEnv<'w>,
         From,
         (),
         Payment,
         Gas,
-        DeployCall<InteractorEnvExec<'w>, Code<CodeValue>>,
+        DeployCall<InteractorExecEnv<'w>, Code<CodeValue>>,
         RH,
     >
 where
-    From: TxFromSpecified<InteractorEnvExec<'w>>,
-    Payment: TxPayment<InteractorEnvExec<'w>>,
-    Gas: TxGas<InteractorEnvExec<'w>>,
-    CodeValue: TxCodeValue<InteractorEnvExec<'w>>,
-    RH: RHListExec<TxResponse, InteractorEnvExec<'w>>,
+    From: TxFromSpecified<InteractorExecEnv<'w>>,
+    Payment: TxPayment<InteractorExecEnv<'w>>,
+    Gas: TxGas<InteractorExecEnv<'w>>,
+    CodeValue: TxCodeValue<InteractorExecEnv<'w>>,
+    RH: RHListExec<TxResponse, InteractorExecEnv<'w>>,
     RH::ListReturns: NestedTupleFlatten,
 {
-    type Exec = InteractorDeployStep<'w, RH>;
+    type Exec = InteractorExecStep<'w, ScDeployStep, RH>;
 
     fn prepare_async(self) -> Self::Exec {
-        InteractorDeployStep {
+        InteractorExecStep {
             step_wrapper: self.tx_to_step(),
         }
     }
 }
 
-impl<'w, RH> InteractorDeployStep<'w, RH>
+impl<'w, RH> InteractorExecStep<'w, ScDeployStep, RH>
 where
-    RH: RHListExec<TxResponse, InteractorEnvExec<'w>>,
+    RH: RHListExec<TxResponse, InteractorExecEnv<'w>>,
     RH::ListReturns: NestedTupleFlatten,
 {
     pub async fn run(mut self) -> <RH::ListReturns as NestedTupleFlatten>::Unpacked {
@@ -94,7 +82,7 @@ impl Interactor {
         let tx = f(tx_base);
 
         let mut step_wrapper = tx.tx_to_step();
-        self.sc_deploy(&mut step_wrapper.step);
+        self.sc_deploy(&mut step_wrapper.step).await;
         step_wrapper.process_result();
 
         self
@@ -128,7 +116,7 @@ impl Interactor {
         let tx = f(tx_base);
 
         let mut step_wrapper = tx.tx_to_step();
-        self.sc_deploy(&mut step_wrapper.step);
+        self.sc_deploy(&mut step_wrapper.step).await;
         step_wrapper.process_result()
     }
 }
