--- framework/snippets/src/itx.rs
@@ -1,13 +1,0 @@
-#![allow(unused)] // TEMP
-
-mod interactor_env;
-mod interactor_env_deploy;
-mod interactor_env_exec;
-mod interactor_env_query;
-mod interactor_env_transf;
-
-pub use interactor_env::*;
-pub use interactor_env_deploy::*;
-pub use interactor_env_exec::InteractorEnvExec;
-pub use interactor_env_query::InteractorEnvQuery;
-pub use interactor_env_transf::*;
