--- framework/snippets/src/itx/interactor_env_exec.rs
@@ -1,121 +1,0 @@
-use std::path::PathBuf;
-
-use multiversx_sc_scenario::{
-    api::StaticApi,
-    multiversx_sc::{
-        tuple_util::NestedTupleFlatten,
-        types::{
-            AnnotatedValue, Code, DeployCall, FunctionCall, ManagedAddress, ManagedBuffer,
-            RHListExec, Tx, TxBaseWithEnv, TxCodeSource, TxCodeSourceSpecified, TxCodeValue, TxEnv,
-            TxFromSpecified, TxGas, TxPayment, TxToSpecified,
-        },
-    },
-    scenario::tx_to_step::{StepWrapper, TxToStep},
-    scenario_model::{AddressValue, BytesValue, ScCallStep, ScDeployStep, TxExpect, TxResponse},
-    ScenarioTxEnv, ScenarioTxEnvData, ScenarioTxRun, ScenarioWorld,
-};
-
-use crate::{Interactor, InteractorPrepareAsync};
-
-/// Environment for executing transactions.
-pub struct InteractorEnvExec<'w> {
-    pub world: &'w mut Interactor,
-    pub data: ScenarioTxEnvData,
-}
-
-impl<'w> TxEnv for InteractorEnvExec<'w> {
-    type Api = StaticApi;
-
-    type RHExpect = TxExpect;
-
-    fn resolve_sender_address(&self) -> ManagedAddress<Self::Api> {
-        panic!("Explicit sender address expected")
-    }
-
-    fn default_gas_annotation(&self) -> ManagedBuffer<Self::Api> {
-        self.data.default_gas_annotation()
-    }
-
-    fn default_gas_value(&self) -> u64 {
-        self.data.default_gas_value()
-    }
-}
-
-impl<'w> ScenarioTxEnv for InteractorEnvExec<'w> {
-    fn env_data(&self) -> &ScenarioTxEnvData {
-        &self.data
-    }
-}
-
-pub struct InteractorCallStep<'w, RH>
-where
-    RH: RHListExec<TxResponse, InteractorEnvExec<'w>>,
-    RH::ListReturns: NestedTupleFlatten,
-{
-    step_wrapper: StepWrapper<InteractorEnvExec<'w>, ScCallStep, RH>,
-}
-
-impl<'w, From, To, Payment, Gas, RH> InteractorPrepareAsync
-    for Tx<InteractorEnvExec<'w>, From, To, Payment, Gas, FunctionCall<StaticApi>, RH>
-where
-    From: TxFromSpecified<InteractorEnvExec<'w>>,
-    To: TxToSpecified<InteractorEnvExec<'w>>,
-    Payment: TxPayment<InteractorEnvExec<'w>>,
-    Gas: TxGas<InteractorEnvExec<'w>>,
-    RH: RHListExec<TxResponse, InteractorEnvExec<'w>>,
-    RH::ListReturns: NestedTupleFlatten,
-{
-    type Exec = InteractorCallStep<'w, RH>;
-
-    fn prepare_async(self) -> Self::Exec {
-        InteractorCallStep {
-            step_wrapper: self.tx_to_step(),
-        }
-    }
-}
-
-impl<'w, RH> InteractorCallStep<'w, RH>
-where
-    RH: RHListExec<TxResponse, InteractorEnvExec<'w>>,
-    RH::ListReturns: NestedTupleFlatten,
-{
-    pub async fn run(mut self) -> <RH::ListReturns as NestedTupleFlatten>::Unpacked {
-        self.step_wrapper
-            .env
-            .world
-            .sc_call(&mut self.step_wrapper.step)
-            .await;
-        self.step_wrapper.process_result()
-    }
-}
-
-impl Interactor {
-    pub fn tx(&mut self) -> TxBaseWithEnv<InteractorEnvExec<'_>> {
-        let data = self.new_env_data();
-        let env = InteractorEnvExec { world: self, data };
-        Tx::new_with_env(env)
-    }
-
-    pub async fn chain_call<From, To, Payment, Gas, RH, F>(&mut self, f: F) -> &mut Self
-    where
-        From: TxFromSpecified<ScenarioTxEnvData>,
-        To: TxToSpecified<ScenarioTxEnvData>,
-        Payment: TxPayment<ScenarioTxEnvData>,
-        Gas: TxGas<ScenarioTxEnvData>,
-        RH: RHListExec<TxResponse, ScenarioTxEnvData, ListReturns = ()>,
-        F: FnOnce(
-            TxBaseWithEnv<ScenarioTxEnvData>,
-        )
-            -> Tx<ScenarioTxEnvData, From, To, Payment, Gas, FunctionCall<StaticApi>, RH>,
-    {
-        let env = self.new_env_data();
-        let tx_base = TxBaseWithEnv::new_with_env(env);
-        let tx = f(tx_base);
-
-        let mut step_wrapper = tx.tx_to_step();
-        self.sc_call(&mut step_wrapper.step).await;
-        step_wrapper.process_result();
-        
-        self
-    }
-}
