+++ framework/snippets/src/lib.rs
@@ -8,9 +8,9 @@ mod interactor_sc_deploy;
 mod interactor_sc_extra;
 mod interactor_sc_transfer;
 mod interactor_sender;
+mod interactor_tx;
 mod interactor_tx_spec;
 mod interactor_vm_query;
-mod itx;
 mod step_buffer;
 
 pub use env_logger;
@@ -18,8 +18,8 @@ pub use hex;
 pub use interactor::*;
 pub use interactor_dns::*;
 pub use interactor_sender::*;
+pub use interactor_tx::*;
 pub use interactor_tx_spec::*;
-pub use itx::*;
 pub use log;
 pub use multiversx_sc_scenario::{self, multiversx_sc};
 pub use multiversx_sdk as erdrs; // TODO: remove
